Split panel into deploying and steady components
