+++ ui/app/components/job-status/panel.hbs
@@ -1,117 +1,5 @@
-<div class="job-status-panel boxed-section {{if this.isActivelyDeploying "active-deployment is-info"}}" data-test-job-status-panel data-test-status-mode={{@statusMode}}>
-  <div class="boxed-section-head">
-    {{#if this.isActivelyDeploying}}
-      <div class="boxed-section-row"
-      {{did-insert (action this.establishOldAllocBlockIDs)}}
-      >
-        Deployment Status
-        <span class="badge is-white bumper-left" data-test-active-deployment-stat="id">{{@job.latestDeployment.shortId}}</span>
-        <div class="pull-right">
-          {{#if @job.latestDeployment.isRunning}}
-            <TwoStepButton
-              data-test-fail
-              @classes={{hash
-                idleButton="is-danger"
-                confirmationMessage="inherit-color"
-                confirmButton="is-danger"}}
-              @idleText="Fail Deployment"
-              @cancelText="Cancel"
-              @confirmText="Yes, Fail Deployment"
-              @confirmationMessage="Are you sure?"
-              @inlineText={{true}}
-              @awaitingConfirmation={{this.fail.isRunning}}
-              @disabled={{this.fail.isRunning}}
-              @onConfirm={{perform this.fail}} />
-          {{/if}}
-          {{#if @job.latestDeployment.requiresPromotion}}
-            <button
-              data-test-promote-canary
-              type="button"
-              class="button is-warning is-small {{if this.promote.isRunning "is-loading"}}"
-              disabled={{this.promote.isRunning}}
-              onclick={{perform this.promote}}>Promote Canary</button>
-          {{/if}}
-        </div>
-      </div>
-    {{else}}
-      <h2>Status</h2>
-      <div class="select-mode">
-        <button type="button"
-          data-test-status-mode-current
-          class="button is-small is-borderless {{if (eq @statusMode "current") "is-active"}}"
-          {{on "click" (action (optional @setStatusMode) "current")}}
-        >
-          Current
-        </button>
-        <button type="button"
-          data-test-status-mode-historical
-          class="button is-small is-borderless {{if (eq @statusMode "historical") "is-active"}}"
-          {{on "click" (action (optional @setStatusMode) "historical")}}>
-          Historical
-        </button>
-      </div>
-    {{/if}}
-  </div>
-  <div class="boxed-section-body">
-
-    {{#if (eq @statusMode "historical")}}
-      <JobPage::Parts::SummaryChart @job={{@job}} />
-    {{else}}
-      {{#if this.isActivelyDeploying}}
-        {{#if this.oldVersionAllocBlockIDs.length}}
-          <h4 class="title is-5">Previous Allocations: {{#if this.oldVersionAllocBlocks.running}}{{this.oldVersionAllocBlocks.running.length}} running{{/if}}</h4>
-          <JobStatus::AllocationStatusRow @allocBlocks={{this.oldVersionAllocBlocks}} />
-        {{/if}}
-
-        <h4 class="title is-5">New allocations: {{this.newVersionAllocBlocks.running.length}}/{{this.totalAllocs}} running</h4>
-        <JobStatus::AllocationStatusRow @allocBlocks={{this.newVersionAllocBlocks}} />
-      {{else}}
-        <h3 class="title is-4 running-allocs-title"><strong>{{@job.runningAllocs}}/{{this.totalAllocs}}</strong> Allocations Running</h3>
-
-        <JobStatus::AllocationStatusRow @allocBlocks={{this.allocBlocks}} />
-      {{/if}}
-
-
-      <div class="sub-panels">
-          {{#if this.isActivelyDeploying}}
-            <legend>
-                {{#each this.allocTypes as |type|}}
-                  <span class="legend-item {{if (eq (get (get this.newVersionAllocBlocks type.label) "length") 0) "faded"}}">
-                    <span class="represented-allocation {{type.label}}"></span>
-                    {{get (get this.newVersionAllocBlocks type.label) "length"}} {{capitalize type.label}}
-                  </span>
-                {{/each}}
-            </legend>
-          {{else}}
-            <legend>
-              {{#each this.allocTypes as |type|}}
-                <span class="legend-item {{if (eq (get (get this.allocBlocks type.label) "length") 0) "faded"}}">
-                  <span class="represented-allocation {{type.label}}"></span>
-                  {{get (get this.allocBlocks type.label) "length"}} {{capitalize type.label}}
-                </span>
-              {{/each}}
-            </legend>
-
-            <section class="versions">
-              <h4>Versions</h4>
-              <ul>
-                {{#each-in this.versions as |version allocs|}}
-                  <li>
-                    <Hds::Badge @text={{version}} />
-                    <Hds::BadgeCount @text={{allocs.length}} @type="inverted" />
-                  </li>
-                {{/each-in}}
-              </ul>
-            </section>
-          {{/if}}
-      </div>
-
-      {{#if this.isActivelyDeploying}}
-        <div class="active-deployment-info-panels">
-          <JobStatus::DeploymentHistory @deployment={{@job.latestDeployment}} />
-          <JobStatus::UpdateParams @job={{@job}} />
-        </div>
-      {{/if}}
-    {{/if}}
-  </div>
-</div>
+{{#if this.isActivelyDeploying}}
+  <JobStatus::Panel::Deploying @job={{@job}} />
+{{else}}
+  <JobStatus::Panel::Steady @job={{@job}} @statusMode={{@statusMode}} @setStatusMode={{@setStatusMode}} />
+{{/if}}
