+++ ui/app/components/job-status/panel/deploying.hbs
@@ -1,0 +1,62 @@
+<div class="job-status-panel boxed-section active-deployment is-info" data-test-job-status-panel>
+  <div class="boxed-section-head">
+    <div class="boxed-section-row"
+      {{did-insert (action this.establishOldAllocBlockIDs)}}
+      >
+        Deployment Status
+        <span class="badge is-white bumper-left" data-test-active-deployment-stat="id">{{@job.latestDeployment.shortId}}</span>
+        <div class="pull-right">
+          {{#if @job.latestDeployment.isRunning}}
+            <TwoStepButton
+              data-test-fail
+              @classes={{hash
+                idleButton="is-danger"
+                confirmationMessage="inherit-color"
+                confirmButton="is-danger"}}
+              @idleText="Fail Deployment"
+              @cancelText="Cancel"
+              @confirmText="Yes, Fail Deployment"
+              @confirmationMessage="Are you sure?"
+              @inlineText={{true}}
+              @awaitingConfirmation={{this.fail.isRunning}}
+              @disabled={{this.fail.isRunning}}
+              @onConfirm={{perform this.fail}} />
+          {{/if}}
+          {{#if @job.latestDeployment.requiresPromotion}}
+            <button
+              data-test-promote-canary
+              type="button"
+              class="button is-warning is-small {{if this.promote.isRunning "is-loading"}}"
+              disabled={{this.promote.isRunning}}
+              onclick={{perform this.promote}}>Promote Canary</button>
+          {{/if}}
+        </div>
+      </div>
+    </div>
+    <div class="boxed-section-body">
+      {{#if this.oldVersionAllocBlockIDs.length}}
+        <h4 class="title is-5">Previous Allocations: {{#if this.oldVersionAllocBlocks.running}}{{this.oldVersionAllocBlocks.running.length}} running{{/if}}</h4>
+        <JobStatus::AllocationStatusRow @allocBlocks={{this.oldVersionAllocBlocks}} />
+      {{/if}}
+
+      <h4 class="title is-5">New allocations: {{this.newVersionAllocBlocks.running.length}}/{{this.totalAllocs}} running</h4>
+      <JobStatus::AllocationStatusRow @allocBlocks={{this.newVersionAllocBlocks}} />
+
+      <div class="legend-and-summary">
+        <legend>
+          {{#each this.allocTypes as |type|}}
+            <span class="legend-item {{if (eq (get (get this.newVersionAllocBlocks type.label) "length") 0) "faded"}}">
+              <span class="represented-allocation {{type.label}}"></span>
+              {{get (get this.newVersionAllocBlocks type.label) "length"}} {{capitalize type.label}}
+            </span>
+          {{/each}}
+        </legend>
+      </div>
+
+      <div class="active-deployment-info-panels">
+        <JobStatus::DeploymentHistory @deployment={{@job.latestDeployment}} />
+        <JobStatus::UpdateParams @job={{@job}} />
+      </div>
+
+    </div>
+</div>
