+++ ui/app/components/job-status/panel/steady.hbs
@@ -1,0 +1,52 @@
+<div class="job-status-panel boxed-section" data-test-job-status-panel data-test-status-mode={{@statusMode}}>
+  <div class="boxed-section-head">
+    <h2>Status</h2>
+    <div class="select-mode">
+      <button type="button"
+        data-test-status-mode-current
+        class="button is-small is-borderless {{if (eq @statusMode "current") "is-active"}}"
+        {{on "click" (action (optional @setStatusMode) "current")}}
+      >
+        Current
+      </button>
+      <button type="button"
+        data-test-status-mode-historical
+        class="button is-small is-borderless {{if (eq @statusMode "historical") "is-active"}}"
+        {{on "click" (action (optional @setStatusMode) "historical")}}>
+        Historical
+      </button>
+    </div>
+  </div>
+  <div class="boxed-section-body">
+    {{#if (eq @statusMode "historical")}}
+      <JobPage::Parts::SummaryChart @job={{@job}} />
+    {{else}}
+      <h3 class="title is-4 running-allocs-title"><strong>{{@job.runningAllocs}}/{{this.totalAllocs}}</strong> Allocations Running</h3>
+      <JobStatus::AllocationStatusRow @allocBlocks={{this.allocBlocks}} />
+
+      <div class="legend-and-summary">
+        <legend>
+          {{#each this.allocTypes as |type|}}
+            <span class="legend-item {{if (eq (get (get this.allocBlocks type.label) "length") 0) "faded"}}">
+              <span class="represented-allocation {{type.label}}"></span>
+              {{get (get this.allocBlocks type.label) "length"}} {{capitalize type.label}}
+            </span>
+          {{/each}}
+        </legend>
+
+        <section class="versions">
+          <h4>Versions</h4>
+          <ul>
+            {{#each-in this.versions as |version allocs|}}
+              <li>
+                <Hds::Badge @text={{version}} />
+                <Hds::BadgeCount @text={{allocs.length}} @type="inverted" />
+              </li>
+            {{/each-in}}
+          </ul>
+        </section>
+
+      </div>
+    {{/if}}
+  </div>
+</div>
