+++ ui/app/styles/components/job-status-panel.scss
@@ -128,7 +128,7 @@
     }
   }
 
-  & > .boxed-section-body > .sub-panels {
+  & > .boxed-section-body > .legend-and-summary {
     display: grid;
     gap: 0.5rem;
     grid-template-columns: 50% 50%;
